refactor: move slashcommands modules
