--- apps/meteor/app/slashcommand-asciiarts/client/index.ts
@@ -1,5 +1,0 @@
-import './gimme';
-import './lenny';
-import './shrug';
-import './tableflip';
-import './unflip';
