--- apps/meteor/app/slashcommands-archiveroom/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
