--- apps/meteor/app/slashcommands-ban/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
