--- apps/meteor/app/slashcommands-create/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
