--- apps/meteor/app/slashcommands-hide/client/index.ts
@@ -1,1 +1,0 @@
-import './hide';
