--- apps/meteor/app/slashcommands-invite/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
