--- apps/meteor/app/slashcommands-inviteall/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
