--- apps/meteor/app/slashcommands-join/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
