--- apps/meteor/app/slashcommands-kick/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
