--- apps/meteor/app/slashcommands-open/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
