--- apps/meteor/app/slashcommands-status/client/index.ts
@@ -1,1 +1,0 @@
-import './status';
