--- apps/meteor/app/slashcommands-topic/client/index.ts
@@ -1,1 +1,0 @@
-import './topic';
