--- apps/meteor/app/slashcommands-unarchiveroom/client/index.ts
@@ -1,1 +1,0 @@
-import './client';
