+++ apps/meteor/client/importPackages.ts
@@ -1,15 +1,3 @@
 import '../app/lib/client';
-import '../app/slashcommands-archiveroom/client';
-import '../app/slashcommand-asciiarts/client';
-import '../app/slashcommands-ban/client';
-import '../app/slashcommands-create/client';
-import '../app/slashcommands-hide/client';
-import '../app/slashcommands-invite/client';
-import '../app/slashcommands-inviteall/client';
-import '../app/slashcommands-join/client';
-import '../app/slashcommands-kick/client';
-import '../app/slashcommands-open/client';
-import '../app/slashcommands-topic/client';
-import '../app/slashcommands-unarchiveroom/client';
 import '../app/utils/client';
 import '../app/ui-utils/client';
