+++ apps/meteor/client/startup/slashCommands/index.ts
@@ -1,2 +1,18 @@
+import './archive';
+import './ban';
+import './create';
 import './federation';
+import './gimme';
+import './hide';
+import './invite';
+import './inviteAllTo';
+import './join';
+import './kick';
+import './lenny';
+import './open';
+import './shrug';
 import './slackbridgeImport';
+import './tableflip';
+import './topic';
+import './unarchive';
+import './unflip';
