+++ apps/meteor/client/startup/slashCommands/inviteAllTo.ts
@@ -1,6 +1,6 @@
-import { queryClient } from '../../../client/lib/queryClient';
-import { roomsQueryKeys } from '../../../client/lib/queryKeys';
-import { slashCommands } from '../../utils/client/slashCommand';
+import { slashCommands } from '../../../app/utils/client/slashCommand';
+import { queryClient } from '../../lib/queryClient';
+import { roomsQueryKeys } from '../../lib/queryKeys';
 
 const invalidateMembers = (_err: unknown, _result: unknown, params: { msg: { rid: string } }) => {
 	if (!_err) void queryClient.invalidateQueries({ queryKey: [...roomsQueryKeys.room(params.msg.rid), 'members'] });
