+++ apps/meteor/client/startup/slashCommands/kick.ts
@@ -1,8 +1,8 @@
 import type { SlashCommandCallbackParams } from '@rocket.chat/core-typings';
 
-import { queryClient } from '../../../client/lib/queryClient';
-import { roomsQueryKeys } from '../../../client/lib/queryKeys';
-import { slashCommands } from '../../utils/client/slashCommand';
+import { slashCommands } from '../../../app/utils/client/slashCommand';
+import { queryClient } from '../../lib/queryClient';
+import { roomsQueryKeys } from '../../lib/queryKeys';
 
 slashCommands.add({
 	command: 'kick',
