+++ apps/meteor/client/startup/slashCommands/shrug.ts
@@ -1,7 +1,7 @@
 import type { SlashCommandCallbackParams } from '@rocket.chat/core-typings';
 
-import { sdk } from '../../utils/client/lib/SDKClient';
-import { slashCommands } from '../../utils/client/slashCommand';
+import { sdk } from '../../../app/utils/client/lib/SDKClient';
+import { slashCommands } from '../../../app/utils/client/slashCommand';
 /*
  * Shrug is a named function that will replace /shrug commands
  * @param {Object} message - The message object
