+++ apps/meteor/client/startup/slashCommands/status.ts
@@ -1,8 +1,8 @@
 import type { SlashCommandCallbackParams } from '@rocket.chat/core-typings';
 
-import { dispatchToastMessage } from '../../../client/lib/toast';
-import { sdk } from '../../utils/client/lib/SDKClient';
-import { slashCommands } from '../../utils/client/slashCommand';
+import { sdk } from '../../../app/utils/client/lib/SDKClient';
+import { slashCommands } from '../../../app/utils/client/slashCommand';
+import { dispatchToastMessage } from '../../lib/toast';
 
 slashCommands.add({
 	command: 'status',
