+++ apps/meteor/client/startup/slashCommands/topic.ts
@@ -1,11 +1,11 @@
 import type { SlashCommandCallbackParams } from '@rocket.chat/core-typings';
 import { clientCallbacks } from '@rocket.chat/ui-client';
 
-import { hasPermission } from '../../../client/lib/authorization';
-import { dispatchToastMessage } from '../../../client/lib/toast';
-import { Rooms } from '../../../client/stores';
-import { sdk } from '../../utils/client/lib/SDKClient';
-import { slashCommands } from '../../utils/client/slashCommand';
+import { sdk } from '../../../app/utils/client/lib/SDKClient';
+import { slashCommands } from '../../../app/utils/client/slashCommand';
+import { hasPermission } from '../../lib/authorization';
+import { dispatchToastMessage } from '../../lib/toast';
+import { Rooms } from '../../stores';
 
 slashCommands.add({
 	command: 'topic',
